requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiojobs==1.1.0
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.5.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
environs==9.5.0
et-xmlfile==1.1.0
frozenlist==1.4.0
humanfriendly==10.0
idna==3.4
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
openpyxl==3.1.2
packaging==23.1
#pandas==2.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
structlog==23.1.0
typing_extensions==4.10.0
tzdata==2024.1
yarl==1.9.2
PyYAML==6.0.1